% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{efficiency}
\alias{efficiency}
\title{Function to Calculate Efficiency (D- AND A-EFFICIENCY)}
\usage{
efficiency(data, variables, labels, corrvar)
}
\arguments{
\item{data}{- input data}

\item{variables}{- list of variables to include in the design matrix (should match with the column names of data)}

\item{labels}{- list of labels for variables}

\item{corrvar}{- name of the column corresponding to correlation coefficients in the input data}
}
\value{
-  Number of studies, a_efficiency, d_efficiency
}
\description{
Function to Calculate Efficiency (D- AND A-EFFICIENCY)
}
