% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency_by_adding_one_full_fact.R
\name{efficiency_by_adding_one_full_fact}
\alias{efficiency_by_adding_one_full_fact}
\title{FUNCTION TO CALCULATE EFFICIENCY AFTER ADDING A NEW STUDY}
\usage{
efficiency_by_adding_one_full_fact(data, variables, labels, corrvar, naval)
}
\arguments{
\item{data}{- input data}

\item{variables}{- list of variables to include in the design matrix (should match with the column names of data)}

\item{labels}{- list of labels for variables}

\item{corrvar}{- name of the column corresponding to correlation coefficients}

\item{naval}{- value of variable indicating the variable cannot be identified in the study}
}
\value{
-  a_efficiency, d_efficiency
}
\description{
FUNCTION TO CALCULATE EFFICIENCY AFTER ADDING A NEW STUDY
}
