% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency_next_best_studies.R
\name{efficiency_next_best_studies}
\alias{efficiency_next_best_studies}
\title{Function to get the "Best Next Studies"}
\usage{
efficiency_next_best_studies(
  data,
  variables,
  labels,
  corrvar,
  naval,
  nstudies_add
)
}
\arguments{
\item{data}{- input data}

\item{variables}{- list of variables to include in the design matrix (should match with the column names of data)}

\item{labels}{- list of labels for variables}

\item{corrvar}{- name of the column corresponding to correlation coefficients}

\item{naval}{- value of variable indicating the variable cannot be identified in the study}

\item{nstudies_add}{- number of next studies}
}
\value{
- List of two data frames (one based on A-efficiency and another based on D-efficiency)
}
\description{
Function to get the "Best Next Studies"
}
