% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency_over_time.R
\name{efficiency_over_time}
\alias{efficiency_over_time}
\title{Function to get efficiency over time}
\usage{
efficiency_over_time(data, variables, labels, corrvar, naval, yrvar, yrs)
}
\arguments{
\item{data}{- input data}

\item{variables}{- list of variables to include in the design matrix (should match with the column names of data)}

\item{labels}{- list of labels for variables}

\item{corrvar}{- name of the column corresponding to correlation coefficients}

\item{naval}{- value of variable indicating the variable cannot be identified in the study}

\item{yrvar}{-  name of the column corresponding to the year}

\item{yrs}{- vector of years to create the block of studies}
}
\value{
- List of two data frames (one for true efficiency and another for optimal efficiency)
}
\description{
Function to get efficiency over time
}
