% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency_over_time_plot.R
\name{efficiency_over_time_plot}
\alias{efficiency_over_time_plot}
\title{Function to plot efficiency over time}
\usage{
efficiency_over_time_plot(data)
}
\arguments{
\item{data}{- Output from 'efficiency_over_time' function}
}
\value{
- Plot of True and Optimal A and D efficiency
}
\description{
Function to plot efficiency over time
}
